requirements.txt
matplotlib==3.3.3
torchvision==0.8.2+cu110
opencv_python==4.4.0.46
psutil==5.7.3
pycocotools==2.0.2
scipy==1.5.4
torch==1.7.1+cu110
tqdm==4.54.1
numpy==1.19.4
keyboard==0.13.5
pywin32==300
ahk==0.11.1
coremltools==4.0
onnx==1.8.0
pandas==1.1.5
Pillow==8.0.1
PyYAML==5.3.1
seaborn==0.11.1
thop==0.0.31-2005241907
wandb==0.10.12
win32gui==221.6
